+++ src/PhpWord/Element/Footer.php
@@ -116,21 +116,6 @@ class Footer extends AbstractContainer
     }
 
     /**
-     * Add textbox element
-     *
-     * @param mixed $style
-     * @return \PhpOffice\PhpWord\Element\TextBox
-     * @todo Merge with the same function on Section
-     */
-    public function addTextBox($style = null)
-    {
-        $textbox = new TextBox($this->getDocPart(), $this->getDocPartId(), $style);
-        $this->addElement($textbox);
-    
-        return $textbox;
-    }
-    
-    /**
      * Add table element
      *
      * @param mixed $style