+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This release changed PHPWord license from LGPL 2.1 to LGPL 3.
 
 - Image: Ability to define relative and absolute positioning - @basjan GH-217
 - Footer: Conform footer with header by adding firstPage, evenPage and by inheritance - @basjan @ivanlanin GH-219
+- TextBox: Ability to add textbox in section, header, and footer - @basjan @ivanlanin GH-228
 
 ### Bugfixes